Use a typeclass to format stuff
